# PERF: Fix groupby skipna performance